requirements.txt
fastapi
gunicorn
uvicorn
jinja2
pytest
requests
httpx
pre-commit
python-dotenv
pydantic_settings
aiofiles
python-multipart
pillow
pytesseract
tesseract
